Move steam-reduction from general to steam
