+++ src/app/calculator/calculator-routing/general.routing.ts
@@ -10,7 +10,6 @@ import { PreAssessmentComponent } from "../utilities/pre-assessment/pre-assessme
 import { UnitConverterComponent } from "../utilities/unit-converter/unit-converter.component";
 import { WaterReductionComponent } from "../utilities/water-reduction/water-reduction.component";
 import { UtilitiesListComponent } from "../utilities/utilities-list/utilities-list.component";
-import { SteamReductionComponent } from "../utilities/steam-reduction/steam-reduction.component";
 import { TankInsulationReductionComponent } from "../utilities/tank-insulation-reduction/tank-insulation-reduction.component";
 
 export const generalRoutes: Routes = [
@@ -57,10 +56,6 @@ export const generalRoutes: Routes = [
     {
         path: 'water-reduction',
         component: WaterReductionComponent
-    },
-    {
-        path: 'steam-reduction',
-        component: SteamReductionComponent
     },
     {
         path: 'tank-insulation-reduction',
