+++ src/app/calculator/steam/steam-list/steam-list.component.html
@@ -3,6 +3,20 @@
         <h2 class="mb-1">Steam Calculators</h2>
         <div class="d-flex calc-item-row w-100">
             <div class="p-2 col-2 d-flex justify-content-center align-items-center">
+                <div class="calc-icon-div d-flex justify-content-center align-items-center">
+                    <img src="assets/images/calculator-icons/utilities-icons/steam-reduction-icon.png" class="calc-img"
+                        [routerLink]="['/calculators/steam-reduction']">
+                </div>
+            </div>
+            <div class="d-flex flex-fill flex-column justify-content-center">
+                <a [routerLink]="['/calculators/steam-reduction']">Steam Reduction</a>
+                <p>
+                    This calculator is used to quantify the energy savings associated with reducing steam use.
+                </p>
+            </div>
+        </div>
+        <div class="d-flex calc-item-row w-100">
+            <div class="p-2 col-2 d-flex justify-content-center align-items-center">
                 <div class="calc-icon-div d-flex justify-content-center align-items-center steam">
                     <img src="assets/images/calculator-icons/steam-icons/steam-properties.png" class="calc-img"
                         [routerLink]="['/calculators/steam-properties']">
