+++ src/app/calculator/steam/steam-reduction/steam-reduction-form/steam-reduction-form.component.css
