+++ src/app/calculator/steam/steam-reduction/steam-reduction-results/steam-reduction-results.component.ts
