+++ src/app/calculator/utilities/utilities.module.ts
@@ -8,7 +8,6 @@ import { Co2SavingsModule } from './co2-savings/co2-savings.module';
 import { ElectricityReductionModule } from './electricity-reduction/electricity-reduction.module';
 import { NaturalGasReductionModule } from './natural-gas-reduction/natural-gas-reduction.module';
 import { WaterReductionModule } from './water-reduction/water-reduction.module';
-import { SteamReductionModule } from './steam-reduction/steam-reduction.module';
 import { CashFlowModule } from './cash-flow/cash-flow.module';
 import { UnitConverterModule } from './unit-converter/unit-converter.module';
 import { PipeInsulationReductionModule } from './pipe-insulation-reduction/pipe-insulation-reduction.module';
@@ -33,7 +32,6 @@ import { RouterModule } from '@angular/router';
         Co2SavingsModule,
         ElectricityReductionModule,
         NaturalGasReductionModule,
-        SteamReductionModule,
         WaterReductionModule,
         PipeInsulationReductionModule,
         TankInsulationReductionModule,
