+++ src/app/dashboard/sidebar/calculator-list/calculator-list.component.html
@@ -63,6 +63,8 @@
 <div routerLinkActive="show-border" [routerLinkActiveOptions]="{exact:true}">
   <span class="sidebar-item" [routerLink]="['/calculators/steam-list']" routerLinkActive="selected">Steam</span>
   <div class="directory-item">
+    <span [routerLink]="['/calculators/steam-reduction']" routerLinkActive="selected" class="sidebar-item">Steam
+      Reduction</span>
     <span [routerLink]="['/calculators/steam-properties']" routerLinkActive="selected" class="sidebar-item">Steam
       Properties</span>
     <span [routerLink]="['/calculators/saturated-properties']" routerLinkActive="selected"
@@ -136,8 +138,6 @@
       class="sidebar-item">Electricity Reduction</span>
     <span [routerLink]="['/calculators/natural-gas-reduction']" routerLinkActive="selected" class="sidebar-item">Natural
       Gas Reduction</span>
-    <span [routerLink]="['/calculators/steam-reduction']" routerLinkActive="selected" class="sidebar-item">Steam
-      Reduction</span>
     <span [routerLink]="['/calculators/water-reduction']" routerLinkActive="selected" class="sidebar-item">Water
       Reduction</span>
     <span [routerLink]="['/calculators/pipe-insulation-reduction']" routerLinkActive="selected"
